server/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
asyncio==3.4.3
boto3==1.34.23
botocore==1.34.23
certifi==2023.11.17
distro==1.9.0
exceptiongroup==1.2.0
ffmpeg-python==0.2.0
filelock==3.13.1
fsspec==2023.12.2
future==0.18.3
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
Jinja2==3.1.3
jmespath==1.0.1
MarkupSafe==2.1.4
mpmath==1.3.0
networkx==3.2.1
openai==1.9.0
pydantic==2.5.3
pydantic_core==2.14.6
python-dateutil==2.8.2
s3transfer==0.10.0
semantic-version==2.10.0
setuptools-rust==1.8.1
six==1.16.0
sniffio==1.3.0
sympy==1.12
tomli==2.0.1
torch==2.1.2
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==1.26.18
whisper==1.1.10
